requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pymongo==4.6.1
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
python-dateutil==2.8.2
httpx[http2]==0.27.0
aio-pika==9.4.0
